requirements.txt
streamlit>=1.39
openai>=1.50.0
python-dotenv>=1.0.1
faiss-cpu>=1.8.0
tiktoken>=0.7.0
pypdf>=5.0.0
